Extract CommandProcessor and ControlCommandProcessor into separate files
Fixes https://github.com/deivid-rodriguez/pry-byebug/issues/22
